# api/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: agrologistics.uex@gmail.com
    name: AgroLogistics
    url: http://spilab.es/
  description: API for AgroLogistics project of Catedra Telefonica UEX 2022
  termsOfService: http://spilab.es/
  title: AgroLogistics
  version: "1.0"
servers:
- url: /
- url: https://raw.githubusercontent.com/
paths:
...
  /producer_transaction:
    delete:
      description: Delete producer transaction
      operationId: deleteProducerTransaction
      parameters:
      - description: Producer transaction ID
        explode: true
        in: query
        name: producerTransactionID
        required: true
        schema:
          format: int32
          type: integer
        style: form
      responses:
        201:
          content:
            text/plain:
              schema:
                example: Created/Updated/Deleted
                title: Created/Updated/Deleted
                type: string
          description: Created/Updated/Deleted
        404:
          content:
            text/plain:
              schema:
                example: Not found
                title: Endpoint not found
                type: string
          description: Not found response
        405:
          content:
            text/plain:
              schema:
                example: The format is not valid
                title: Invalid input
                type: string
          description: Invalid input
      summary: Delete producer transaction
      tags:
      - ProducerTransaction
      x-swagger-router-controller: ProducerTransaction
    get:
      description: Returns all the producer transaction
      operationId: getProducerTransactions
      parameters:
      - description: ID
        explode: true
        in: query
        name: id
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Day
        explode: true
        in: query
        name: day
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Month
        explode: true
        in: query
        name: month
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Year
        explode: true
        in: query
        name: year
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Range (YYYY-MM-dd/YYYY-MM-dd)
        explode: true
        in: query
        name: range
        required: false
        schema:
          type: string
        style: form
      - description: Logistic Center ID
        explode: true
        in: query
        name: logistic_center_id
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Producer ID
        explode: true
        in: query
        name: producer_id
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Product ID
        explode: true
        in: query
        name: product_id
        required: false
        schema:
          format: int32
          type: integer
        style: form
      responses:
        200:
          content:
            application/json:
              schema:
                example: '[{"id":"1", "product_id": "1", "logistic_center_id": "1",
                  "producer_id": "1", "product_category": "high_quality", "amount_kg":
                  "100", "date": "2022-03-10 10:00:00", "price": 10, "storage_type": "automatic"}]'
                title: Info
                type: string
          description: OK
        201:
          content:
            text/plain:
              schema:
                example: Created/Updated/Deleted
                title: Created/Updated/Deleted
                type: string
          description: Created/Updated/Deleted
        404:
          content:
            text/plain:
              schema:
                example: Not found
                title: Endpoint not found
                type: string
          description: Not found response
        405:
          content:
            text/plain:
              schema:
                example: The format is not valid
                title: Invalid input
                type: string
          description: Invalid input
      summary: Returns all the producer transaction
      tags:
      - ProducerTransaction
      x-swagger-router-controller: ProducerTransaction
    post:
      description: Creates new producer transaction
      operationId: postProducerTransaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProducerTransaction'
        required: true
      responses:
        201:
          content:
            text/plain:
              schema:
                example: Created/Updated/Deleted
                title: Created/Updated/Deleted
                type: string
          description: Created/Updated/Deleted
        404:
          content:
            text/plain:
              schema:
                example: Not found
                title: Endpoint not found
                type: string
          description: Not found response
        405:
          content:
            text/plain:
              schema:
                example: The format is not valid
                title: Invalid input
                type: string
          description: Invalid input
      summary: Creates new producer transaction
      tags:
      - ProducerTransaction
      x-swagger-router-controller: ProducerTransaction
    put:
      description: Modifies a producer transaction
      operationId: putProducerTransaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProducerTransaction'
        required: true
      responses:
        201:
          content:
            text/plain:
              schema:
                example: Created/Updated/Deleted
                title: Created/Updated/Deleted
                type: string
          description: Created/Updated/Deleted
        404:
          content:
            text/plain:
              schema:
                example: Not found
                title: Endpoint not found
                type: string
          description: Not found response
        405:
          content:
            text/plain:
              schema:
                example: The format is not valid
                title: Invalid input
                type: string
          description: Invalid input
      summary: Modifies a producer transaction
      tags:
      - ProducerTransaction
      x-swagger-router-controller: ProducerTransaction
...
  /consumer_transaction:
    delete:
      description: Delete consumer transaction
      operationId: deleteConsumerTransaction
      parameters:
      - description: Consumer transaction ID
        explode: true
        in: query
        name: ConsumerTransactionID
        required: true
        schema:
          format: int32
          type: integer
        style: form
      responses:
        201:
          content:
            text/plain:
              schema:
                example: Created/Updated/Deleted
                title: Created/Updated/Deleted
                type: string
          description: Created/Updated/Deleted
        404:
          content:
            text/plain:
              schema:
                example: Not found
                title: Endpoint not found
                type: string
          description: Not found response
        405:
          content:
            text/plain:
              schema:
                example: The format is not valid
                title: Invalid input
                type: string
          description: Invalid input
      summary: Delete consumer transaction
      tags:
      - ConsumerTransaction
      x-swagger-router-controller: ConsumerTransaction
    get:
      description: Returns all the consumer transaction
      operationId: getConsumerTransactions
      parameters:
      - description: ID
        explode: true
        in: query
        name: id
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Day
        explode: true
        in: query
        name: day
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Month
        explode: true
        in: query
        name: month
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Year
        explode: true
        in: query
        name: year
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Range (YYYY-MM-dd/YYYY-MM-dd)
        explode: true
        in: query
        name: range
        required: false
        schema:
          type: string
        style: form
      - description: Logistic Center ID
        explode: true
        in: query
        name: logistic_center_id
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Consumer ID
        explode: true
        in: query
        name: consumer_id
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Product ID
        explode: true
        in: query
        name: product_id
        required: false
        schema:
          format: int32
          type: integer
        style: form
      responses:
        200:
          content:
            application/json:
              schema:
                example: '[{"id":"1", "product_id": "1", "logistic_center_id": "1",
                  "consumer_id": "1", "product_category": "high_quality", "amount_kg":
                  "100", "date": "2022-03-10 10:00:00", "price": 10, "storage_type": "automatic"}]'
                title: Info
                type: string
          description: OK
        201:
          content:
            text/plain:
              schema:
                example: Created/Updated/Deleted
                title: Created/Updated/Deleted
                type: string
          description: Created/Updated/Deleted
        404:
          content:
            text/plain:
              schema:
                example: Not found
                title: Endpoint not found
                type: string
          description: Not found response
        405:
          content:
            text/plain:
              schema:
                example: The format is not valid
                title: Invalid input
                type: string
          description: Invalid input
      summary: Returns all the consumer transaction
      tags:
      - ConsumerTransaction
      x-swagger-router-controller: ConsumerTransaction
    post:
      description: Creates new consumer transaction
      operationId: postConsumerTransaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumerTransaction'
        required: true
      responses:
        201:
          content:
            text/plain:
              schema:
                example: Created/Updated/Deleted
                title: Created/Updated/Deleted
                type: string
          description: Created/Updated/Deleted
        404:
          content:
            text/plain:
              schema:
                example: Not found
                title: Endpoint not found
                type: string
          description: Not found response
        405:
          content:
            text/plain:
              schema:
                example: The format is not valid
                title: Invalid input
                type: string
          description: Invalid input
      summary: Creates new consumer transaction
      tags:
      - ConsumerTransaction
      x-swagger-router-controller: ConsumerTransaction
    put:
      description: Modifies a consumer transaction
      operationId: putConsumerTransaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumerTransaction'
        required: true
      responses:
        201:
          content:
            text/plain:
              schema:
                example: Created/Updated/Deleted
                title: Created/Updated/Deleted
                type: string
          description: Created/Updated/Deleted
        404:
          content:
            text/plain:
              schema:
                example: Not found
                title: Endpoint not found
                type: string
          description: Not found response
        405:
          content:
            text/plain:
              schema:
                example: The format is not valid
                title: Invalid input
                type: string
          description: Invalid input
      summary: Modifies a consumer transaction
      tags:
      - ConsumerTransaction
      x-swagger-router-controller: ConsumerTransaction
...
components:
...
  schemas:
...
    ProducerTransaction:
      example:
        id: 1
        product_id: 1
        product_name: Patata
        logistic_center_id: 1
        logistic_center_name: Caceres Logistic Center
        producer_id: 1
        producer_name: Farmer Lopez
        product_category: high_quality
        amount_kg: 100
        date: "2022-03-10 10:00:00"
        price: 10
        storage_type: automatic
      properties:
        id:
          example: 1
          format: int64
          type: integer
        product_id:
          example: 1
          format: int64
          type: integer
        product_name: 
          example: Patata
          type: string
        logistic_center_id:
          example: 1
          format: int64
          type: integer
        logistic_center_name:
          example: Caceres Logistic Center
          type: string
        producer_id:
          example: 1
          format: int64
          type: integer
        producer_name:
          example: Farmer Lopez
          type: string
        product_category:
          example: high_quality
          type: string
        amount_kg:
          example: 100
          type: number
        date:
          example: "2022-03-10 10:00:00"
          type: string
        price:
          example: 10
          type: number
        storage_type:
          example: automatic
          type: string
      type: object
...
    ConsumerTransaction:
      example:
        id: 1
        product_id: 1
        product_name: Patata
        logistic_center_id: 1
        logistic_center_name: Caceres Logistic Center
        consumer_id: 1
        consumer_name: Caceres Greengrocer
        product_category: high_quality
        amount_kg: 100
        date: "2022-03-10 10:00:00"
        price: 10
        storage_type: automatic
      properties:
        id:
          example: 1
          format: int64
          type: integer
        product_id:
          example: 1
          format: int64
          type: integer
        product_name: 
          example: Patata
          type: string
        logistic_center_id:
          example: 1
          format: int64
          type: integer
        logistic_center_name:
          example: Caceres Logistic Center
          type: string
        consumer_id:
          example: 1
          format: int64
          type: integer
        consumer_name:
          example: Caceres Greengrocer
          type: string
        product_category:
          example: high_quality
          type: string
        amount_kg:
          example: 100
          type: number
        date:
          example: "2022-03-10 10:00:00"
          type: string
        price:
          example: 10
          type: number
        storage_type:
          example: automatic
          type: string
      type: object
...